create helper function for concise code
